chore: rename translator util pkg to translate
