+++ apis/v1alpha2/validation/httproute_test.go
@@ -25,7 +25,7 @@ import (
 	utilpointer "k8s.io/utils/pointer"
 
 	gatewayv1a2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
-	"sigs.k8s.io/gateway-api/apis/v1alpha2/util/translator"
+	"sigs.k8s.io/gateway-api/apis/v1beta1/util/translate"
 )
 
 func TestValidateHTTPRoute(t *testing.T) {
@@ -45,7 +45,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 				Matches: []gatewayv1a2.HTTPRouteMatch{
 					{
 						Path: &gatewayv1a2.HTTPPathMatch{
-							Type:  translator.PathMatchTypePtr("PathPrefix"),
+							Type:  translate.PathMatchTypePtr("PathPrefix"),
 							Value: utilpointer.String("/"),
 						},
 					},
@@ -55,7 +55,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 						BackendRef: gatewayv1a2.BackendRef{
 							BackendObjectReference: gatewayv1a2.BackendObjectReference{
 								Name: testService,
-								Port: translator.PortNumberPtr(8080),
+								Port: translate.PortNumberPtr(8080),
 							},
 							Weight: utilpointer.Int32(100),
 						},
@@ -71,7 +71,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 				Matches: []gatewayv1a2.HTTPRouteMatch{
 					{
 						Path: &gatewayv1a2.HTTPPathMatch{
-							Type:  translator.PathMatchTypePtr("PathPrefix"),
+							Type:  translate.PathMatchTypePtr("PathPrefix"),
 							Value: utilpointer.String("/"),
 						},
 					},
@@ -82,7 +82,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 						RequestMirror: &gatewayv1a2.HTTPRequestMirrorFilter{
 							BackendRef: gatewayv1a2.BackendObjectReference{
 								Name: testService,
-								Port: translator.PortNumberPtr(8081),
+								Port: translate.PortNumberPtr(8081),
 							},
 						},
 					},
@@ -97,7 +97,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 				Matches: []gatewayv1a2.HTTPRouteMatch{
 					{
 						Path: &gatewayv1a2.HTTPPathMatch{
-							Type:  translator.PathMatchTypePtr("PathPrefix"),
+							Type:  translate.PathMatchTypePtr("PathPrefix"),
 							Value: utilpointer.String("/"),
 						},
 					},
@@ -108,7 +108,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 						RequestMirror: &gatewayv1a2.HTTPRequestMirrorFilter{
 							BackendRef: gatewayv1a2.BackendObjectReference{
 								Name: testService,
-								Port: translator.PortNumberPtr(8080),
+								Port: translate.PortNumberPtr(8080),
 							},
 						},
 					},
@@ -117,7 +117,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 						RequestMirror: &gatewayv1a2.HTTPRequestMirrorFilter{
 							BackendRef: gatewayv1a2.BackendObjectReference{
 								Name: specialService,
-								Port: translator.PortNumberPtr(8080),
+								Port: translate.PortNumberPtr(8080),
 							},
 						},
 					},
@@ -132,7 +132,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 				Matches: []gatewayv1a2.HTTPRouteMatch{
 					{
 						Path: &gatewayv1a2.HTTPPathMatch{
-							Type:  translator.PathMatchTypePtr("PathPrefix"),
+							Type:  translate.PathMatchTypePtr("PathPrefix"),
 							Value: utilpointer.String("/"),
 						},
 					},
@@ -154,7 +154,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 						RequestMirror: &gatewayv1a2.HTTPRequestMirrorFilter{
 							BackendRef: gatewayv1a2.BackendObjectReference{
 								Name: testService,
-								Port: translator.PortNumberPtr(8080),
+								Port: translate.PortNumberPtr(8080),
 							},
 						},
 					},
@@ -180,7 +180,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 				Matches: []gatewayv1a2.HTTPRouteMatch{
 					{
 						Path: &gatewayv1a2.HTTPPathMatch{
-							Type:  translator.PathMatchTypePtr("PathPrefix"),
+							Type:  translate.PathMatchTypePtr("PathPrefix"),
 							Value: utilpointer.String("/"),
 						},
 					},
@@ -191,7 +191,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 						RequestMirror: &gatewayv1a2.HTTPRequestMirrorFilter{
 							BackendRef: gatewayv1a2.BackendObjectReference{
 								Name: testService,
-								Port: translator.PortNumberPtr(8080),
+								Port: translate.PortNumberPtr(8080),
 							},
 						},
 					},
@@ -211,7 +211,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 						RequestMirror: &gatewayv1a2.HTTPRequestMirrorFilter{
 							BackendRef: gatewayv1a2.BackendObjectReference{
 								Name: testService,
-								Port: translator.PortNumberPtr(8080),
+								Port: translate.PortNumberPtr(8080),
 							},
 						},
 					},
@@ -231,7 +231,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 						RequestMirror: &gatewayv1a2.HTTPRequestMirrorFilter{
 							BackendRef: gatewayv1a2.BackendObjectReference{
 								Name: specialService,
-								Port: translator.PortNumberPtr(8080),
+								Port: translate.PortNumberPtr(8080),
 							},
 						},
 					},
@@ -246,7 +246,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 				Matches: []gatewayv1a2.HTTPRouteMatch{
 					{
 						Path: &gatewayv1a2.HTTPPathMatch{
-							Type:  translator.PathMatchTypePtr("PathPrefix"),
+							Type:  translate.PathMatchTypePtr("PathPrefix"),
 							Value: utilpointer.String("/"),
 						},
 					},
@@ -268,7 +268,7 @@ func TestValidateHTTPRoute(t *testing.T) {
 						RequestMirror: &gatewayv1a2.HTTPRequestMirrorFilter{
 							BackendRef: gatewayv1a2.BackendObjectReference{
 								Name: testService,
-								Port: translator.PortNumberPtr(8080),
+								Port: translate.PortNumberPtr(8080),
 							},
 						},
 					},
@@ -453,7 +453,7 @@ func TestValidateHTTPBackendUniqueFilters(t *testing.T) {
 					BackendRef: gatewayv1a2.BackendRef{
 						BackendObjectReference: gatewayv1a2.BackendObjectReference{
 							Name: testService,
-							Port: translator.PortNumberPtr(8080),
+							Port: translate.PortNumberPtr(8080),
 						},
 						Weight: utilpointer.Int32(100),
 					},
@@ -463,7 +463,7 @@ func TestValidateHTTPBackendUniqueFilters(t *testing.T) {
 							RequestMirror: &gatewayv1a2.HTTPRequestMirrorFilter{
 								BackendRef: gatewayv1a2.BackendObjectReference{
 									Name: testService,
-									Port: translator.PortNumberPtr(8080),
+									Port: translate.PortNumberPtr(8080),
 								},
 							},
 						},
@@ -480,7 +480,7 @@ func TestValidateHTTPBackendUniqueFilters(t *testing.T) {
 					BackendRef: gatewayv1a2.BackendRef{
 						BackendObjectReference: gatewayv1a2.BackendObjectReference{
 							Name: testService,
-							Port: translator.PortNumberPtr(8080),
+							Port: translate.PortNumberPtr(8080),
 						},
 					},
 					Filters: []gatewayv1a2.HTTPRouteFilter{
@@ -489,7 +489,7 @@ func TestValidateHTTPBackendUniqueFilters(t *testing.T) {
 							RequestMirror: &gatewayv1a2.HTTPRequestMirrorFilter{
 								BackendRef: gatewayv1a2.BackendObjectReference{
 									Name: testService,
-									Port: translator.PortNumberPtr(8080),
+									Port: translate.PortNumberPtr(8080),
 								},
 							},
 						},
@@ -498,7 +498,7 @@ func TestValidateHTTPBackendUniqueFilters(t *testing.T) {
 							RequestMirror: &gatewayv1a2.HTTPRequestMirrorFilter{
 								BackendRef: gatewayv1a2.BackendObjectReference{
 									Name: specialService,
-									Port: translator.PortNumberPtr(8080),
+									Port: translate.PortNumberPtr(8080),
 								},
 							},
 						},
@@ -527,21 +527,21 @@ func TestValidateHTTPPathMatch(t *testing.T) {
 	}{{
 		name: "invalid httpRoute prefix",
 		path: &gatewayv1a2.HTTPPathMatch{
-			Type:  translator.PathMatchTypePtr("PathPrefix"),
+			Type:  translate.PathMatchTypePtr("PathPrefix"),
 			Value: utilpointer.String("/."),
 		},
 		errCount: 1,
 	}, {
 		name: "invalid httpRoute Exact",
 		path: &gatewayv1a2.HTTPPathMatch{
-			Type:  translator.PathMatchTypePtr("Exact"),
+			Type:  translate.PathMatchTypePtr("Exact"),
 			Value: utilpointer.String("/foo/./bar"),
 		},
 		errCount: 1,
 	}, {
 		name: "invalid httpRoute prefix",
 		path: &gatewayv1a2.HTTPPathMatch{
-			Type:  translator.PathMatchTypePtr("PathPrefix"),
+			Type:  translate.PathMatchTypePtr("PathPrefix"),
 			Value: utilpointer.String("/"),
 		},
 		errCount: 0,
@@ -558,7 +558,7 @@ func TestValidateHTTPPathMatch(t *testing.T) {
 						BackendRef: gatewayv1a2.BackendRef{
 							BackendObjectReference: gatewayv1a2.BackendObjectReference{
 								Name: gatewayv1a2.ObjectName("test"),
-								Port: translator.PortNumberPtr(8080),
+								Port: translate.PortNumberPtr(8080),
 							},
 						},
 					}},
@@ -619,7 +619,7 @@ func TestValidateHTTPHeaderMatches(t *testing.T) {
 						BackendRef: gatewayv1a2.BackendRef{
 							BackendObjectReference: gatewayv1a2.BackendObjectReference{
 								Name: gatewayv1a2.ObjectName("test"),
-								Port: translator.PortNumberPtr(8080),
+								Port: translate.PortNumberPtr(8080),
 							},
 						},
 					}},
@@ -683,7 +683,7 @@ func TestValidateHTTPQueryParamMatches(t *testing.T) {
 						BackendRef: gatewayv1a2.BackendRef{
 							BackendObjectReference: gatewayv1a2.BackendObjectReference{
 								Name: gatewayv1a2.ObjectName("test"),
-								Port: translator.PortNumberPtr(8080),
+								Port: translate.PortNumberPtr(8080),
 							},
 						},
 					}},
@@ -840,7 +840,7 @@ func TestValidateHTTPRouteTypeMatchesField(t *testing.T) {
 				Kind:      new(gatewayv1a2.Kind),
 				Name:      "name",
 				Namespace: new(gatewayv1a2.Namespace),
-				Port:      translator.PortNumberPtr(22),
+				Port:      translate.PortNumberPtr(22),
 			}},
 		},
 		errCount: 0,
@@ -946,7 +946,7 @@ func TestValidateHTTPRouteTypeMatchesField(t *testing.T) {
 							BackendRef: gatewayv1a2.BackendRef{
 								BackendObjectReference: gatewayv1a2.BackendObjectReference{
 									Name: gatewayv1a2.ObjectName("test"),
-									Port: translator.PortNumberPtr(8080),
+									Port: translate.PortNumberPtr(8080),
 								},
 							},
 						}},
