+++ apis/v1beta1/util/translate/httproute_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package translator
+package translate
 
 import (
 	"testing"
